+++ bpmnModeler.ts
@@ -1,5 +1,7 @@
 import Modeler from "bpmn-js/lib/Modeler";
-import {TextFileView} from "obsidian";
+import {BpmnPropertiesPanelModule, BpmnPropertiesProviderModule} from 'bpmn-js-properties-panel';
+import {TextFileView, WorkspaceLeaf} from "obsidian";
+import {ObsidianBpmnPluginSettings} from "./settings";
 
 export const VIEW_TYPE_BPMN = "bpmn-view";
 
@@ -8,6 +10,13 @@ export class BpmnModelerView extends TextFileView {
     bpmnDiv: HTMLElement;
     // @ts-ignore
     bpmnModeler: Modeler;
+
+    constructor(
+        public leaf: WorkspaceLeaf,
+        public settings: ObsidianBpmnPluginSettings,
+    ) {
+        super(leaf);
+    }
 
     getViewData() {
         return this.data;
@@ -26,14 +35,26 @@ export class BpmnModelerView extends TextFileView {
         let bpmnSave = this.contentEl.createEl("button", {text: "Save"});
         let bpmnUndo = this.contentEl.createEl("button", {text: "Undo"});
         let bpmnRedo = this.contentEl.createEl("button", {text: "Redo"});
+        let bpmnProperties = this.contentEl.createEl("button", {text: "Properties"});
         this.bpmnDiv = this.contentEl.createEl("div", {cls: "bpmn-view bpmn-fullscreen"});
+        let propertyPanel = this.contentEl.createEl("div", {cls: "bpmn-properties-panel hide"});
         this.bpmnModeler = new Modeler({
             container: this.bpmnDiv,
             keyboard: {
                 bindTo: this.bpmnDiv.win
-            }
+            },
+            propertiesPanel: {
+                parent: propertyPanel
+            },
+            additionalModules: [
+                BpmnPropertiesPanelModule,
+                BpmnPropertiesProviderModule
+                // TODO: Add camunda properties (if requested?) https://github.com/bpmn-io/bpmn-js-examples/tree/master/properties-panel
+            ]
         });
-        this.bpmnDiv.addClass("bpmn-view-white-background");
+        if (this.settings.force_white_background_by_default) {
+            this.bpmnDiv.addClass("bpmn-view-white-background");
+        }
 
         const bpmnModeler = this.bpmnModeler
         const thisRef = this;
@@ -51,6 +72,9 @@ export class BpmnModelerView extends TextFileView {
         });
         bpmnRedo.addEventListener("click", function (e: Event) {
             bpmnModeler.get("commandStack").redo();
+        });
+        bpmnProperties.addEventListener("click", function (e: Event) {
+            propertyPanel.classList.toggle("hide");
         });
     }