+++ typings/index.d.ts
@@ -1,3 +1,4 @@
 declare module 'bpmn-js';
 declare module 'bpmn-js/lib/Modeler';
 declare module 'bpmn-js/lib/NavigatedViewer';
+declare module 'bpmn-js-properties-panel';